New file: server_install.ts
