+++ src/server_manager/web_app/shadowbox_server.ts
@@ -302,36 +302,3 @@ export class ShadowboxServer implements server.Server {
     }
   }
 }
-
-/** Represents the settings needed to launch a dockerized shadowbox. */
-export interface ShadowboxSettings {
-  imageId: string;
-  metricsUrl: string;
-  sentryApiUrl?: string;
-  watchtowerRefreshSeconds?: number;
-}
-
-export function getShellExportCommands(settings: ShadowboxSettings, serverName: string) : string {
-  const variables: {[name: string]: string|number} = {
-    'SB_IMAGE': settings.imageId,
-    'WATCHTOWER_REFRESH_SECONDS': settings.watchtowerRefreshSeconds,
-    'SENTRY_API_URL': settings.sentryApiUrl,
-    'SB_METRICS_URL': settings.metricsUrl,
-  };
-  const lines: string[] = [];
-  for (const name in variables) {
-    if (variables[name]) {
-      lines.push(`export ${name}='${variables[name]}'`);
-    }
-  }
-  lines.push(`export SB_DEFAULT_SERVER_NAME="$(printf '${bashEscape(serverName)}')"`);
-  return lines.join('\n') + '\n';
-}
-
-function bashEscape(s: string): string {
-  // Replace each non-ASCII character with a unicode escape sequence that
-  // is understood by bash.  This avoids an apparent bug in DigitalOcean's
-  // handling of unicode characters in the user_data value.
-  return s.replace(/\P{ASCII}/gu,
-      c => '\\u' + c.charCodeAt(0).toString(16).padStart(4, '0'));
-}
